ui/rsl: split group view into tabs
